deps: update wasi
