+++ packages/preview2-shim/lib/browser/cli.js
@@ -23,12 +23,6 @@ export const exit = {
   }
 };
 
-export const preopens = {
-  getDirectories () {
-    return [];
-  }
-}
-
 export const stdin = {
   getStdin () {
     return 0;
@@ -43,6 +37,36 @@ export const stdout = {
 
 export const stderr = {
   getStderr () {
+    return 2;
+  }
+};
+
+export const terminalInput = {
+  dropTerminalInput () {
+
+  }
+};
+
+export const terminalOutput = {
+  dropTerminalOutput () {
+
+  }
+};
+
+export const terminalStderr = {
+  getTerminalStderr () {
+    return 0;
+  }
+};
+
+export const terminalStdin = {
+  getTerminalStdin () {
+    return 1;
+  }
+};
+
+export const terminalStdout = {
+  getTerminalStdout () {
     return 2;
   }
 };
