--- packages/preview2-shim/types/exports/wasi-http-incoming-handler.d.ts
@@ -1,7 +1,0 @@
-export namespace WasiHttpIncomingHandler {
-  export function handle(request: IncomingRequest, responseOut: ResponseOutparam): void;
-}
-import type { IncomingRequest } from '../exports/wasi-http-types';
-export { IncomingRequest };
-import type { ResponseOutparam } from '../exports/wasi-http-types';
-export { ResponseOutparam };
