--- packages/preview2-shim/types/imports/wasi-cli-base-exit.d.ts
@@ -1,7 +1,0 @@
-export namespace WasiCliBaseExit {
-  /**
-   * Exit the curerent instance and any linked instances.
-   */
-  export function exit(status: Result<void, void>): void;
-}
-export type Result<T, E> = { tag: 'ok', val: T } | { tag: 'err', val: E };
