--- packages/preview2-shim/types/imports/wasi-cli-base-stderr.d.ts
@@ -1,5 +1,0 @@
-export namespace WasiCliBaseStderr {
-  export function getStderr(): OutputStream;
-}
-import type { OutputStream } from '../imports/wasi-io-streams';
-export { OutputStream };
