--- packages/preview2-shim/types/imports/wasi-cli-base-stdin.d.ts
@@ -1,5 +1,0 @@
-export namespace WasiCliBaseStdin {
-  export function getStdin(): InputStream;
-}
-import type { InputStream } from '../imports/wasi-io-streams';
-export { InputStream };
