--- packages/preview2-shim/types/imports/wasi-cli-base-stdout.d.ts
@@ -1,5 +1,0 @@
-export namespace WasiCliBaseStdout {
-  export function getStdout(): OutputStream;
-}
-import type { OutputStream } from '../imports/wasi-io-streams';
-export { OutputStream };
