--- packages/preview2-shim/types/imports/wasi-http-outgoing-handler.d.ts
@@ -1,9 +1,0 @@
-export namespace WasiHttpOutgoingHandler {
-  export function handle(request: OutgoingRequest, options: RequestOptions | null): FutureIncomingResponse;
-}
-import type { OutgoingRequest } from '../imports/wasi-http-types';
-export { OutgoingRequest };
-import type { RequestOptions } from '../imports/wasi-http-types';
-export { RequestOptions };
-import type { FutureIncomingResponse } from '../imports/wasi-http-types';
-export { FutureIncomingResponse };
