--- packages/preview2-shim/types/imports/wasi-sockets-instance-network.d.ts
@@ -1,8 +1,0 @@
-export namespace WasiSocketsInstanceNetwork {
-  /**
-   * Get a handle to the default network.
-   */
-  export function instanceNetwork(): Network;
-}
-import type { Network } from '../imports/wasi-sockets-network';
-export { Network };
