--- packages/preview2-shim/types/wasi-command.d.ts
@@ -1,23 +1,0 @@
-import { WasiCliBaseEnvironment } from './imports/wasi-cli-base-environment';
-import { WasiCliBaseExit } from './imports/wasi-cli-base-exit';
-import { WasiCliBasePreopens } from './imports/wasi-cli-base-preopens';
-import { WasiCliBaseStderr } from './imports/wasi-cli-base-stderr';
-import { WasiCliBaseStdin } from './imports/wasi-cli-base-stdin';
-import { WasiCliBaseStdout } from './imports/wasi-cli-base-stdout';
-import { WasiClocksMonotonicClock } from './imports/wasi-clocks-monotonic-clock';
-import { WasiClocksTimezone } from './imports/wasi-clocks-timezone';
-import { WasiClocksWallClock } from './imports/wasi-clocks-wall-clock';
-import { WasiFilesystemFilesystem } from './imports/wasi-filesystem-filesystem';
-import { WasiIoStreams } from './imports/wasi-io-streams';
-import { WasiPollPoll } from './imports/wasi-poll-poll';
-import { WasiRandomInsecure } from './imports/wasi-random-insecure';
-import { WasiRandomInsecureSeed } from './imports/wasi-random-insecure-seed';
-import { WasiRandomRandom } from './imports/wasi-random-random';
-import { WasiSocketsInstanceNetwork } from './imports/wasi-sockets-instance-network';
-import { WasiSocketsIpNameLookup } from './imports/wasi-sockets-ip-name-lookup';
-import { WasiSocketsNetwork } from './imports/wasi-sockets-network';
-import { WasiSocketsTcp } from './imports/wasi-sockets-tcp';
-import { WasiSocketsTcpCreateSocket } from './imports/wasi-sockets-tcp-create-socket';
-import { WasiSocketsUdp } from './imports/wasi-sockets-udp';
-import { WasiSocketsUdpCreateSocket } from './imports/wasi-sockets-udp-create-socket';
-export function run(): void;
