--- packages/preview2-shim/types/wasi-proxy.d.ts
@@ -1,10 +1,0 @@
-import { WasiHttpOutgoingHandler } from './imports/wasi-http-outgoing-handler';
-import { WasiHttpTypes } from './imports/wasi-http-types';
-import { WasiIoStreams } from './imports/wasi-io-streams';
-import { WasiLoggingHandler } from './imports/wasi-logging-handler';
-import { WasiPollPoll } from './imports/wasi-poll-poll';
-import { WasiRandomInsecure } from './imports/wasi-random-insecure';
-import { WasiRandomInsecureSeed } from './imports/wasi-random-insecure-seed';
-import { WasiRandomRandom } from './imports/wasi-random-random';
-import { WasiHttpIncomingHandler } from './exports/wasi-http-incoming-handler';
-export const incomingHandler: typeof WasiHttpIncomingHandler;
