+++ test/fixtures/wit/deps/cli/stdio.wit
@@ -1,0 +1,17 @@
+interface stdin {
+  use wasi:io/streams.{input-stream}
+
+  get-stdin: func() -> input-stream
+}
+
+interface stdout {
+  use wasi:io/streams.{output-stream}
+
+  get-stdout: func() -> output-stream
+}
+
+interface stderr {
+  use wasi:io/streams.{output-stream}
+
+  get-stderr: func() -> output-stream
+}
