+++ test/fixtures/wit/deps/filesystem/preopens.wit
@@ -1,0 +1,6 @@
+interface preopens {
+    use types.{descriptor}
+
+    /// Return the set of preopened directories, and their path.
+    get-directories: func() -> list<tuple<descriptor, string>>
+}
